requirements.txt
python==3.10.5
fastapi==0.79.1
pydantic==1.9.2
requests==2.28.1
uvicorn==0.17.6
python-dotenv==0.20.0
python-multipart==0.0.5
Jinja2==3.1.2
Markdown==3.4.1
pytest==7.1.2
SQLAlchemy==1.4.40
plotly==5.10.0